+++ packages/deployed-backend-client/src/deployed_backend_client_factory.ts
@@ -62,6 +62,7 @@ export type BackendMetadata = {
     additionalAuthTypes: ApiAuthType[];
     conflictResolutionMode?: ConflictResolutionMode;
     apiId: string;
+    modelSchemaS3Uri: string;
   };
   authConfiguration?: {
     status: BackendDeploymentStatus;