+++ packages/backend-storage/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aws-amplify/backend-storage
 
+## 0.6.0-beta.5
+
+### Patch Changes
+
+- Updated dependencies [5e12247]
+  - @aws-amplify/plugin-types@0.9.0-beta.1
+  - @aws-amplify/backend-output-storage@0.4.0-beta.3
+
 ## 0.6.0-beta.4
 
 ### Patch Changes